requirements.txt
python-dotenv==1.1.0
requests==2.32.3
rich==14.0.0
typer==0.16.0
openai==1.85.0
pydantic==2.11.5
websockets==15.0.1
python-socketio==5.13.0
aiohttp==3.12.11
fastapi==0.115.12
uvicorn==0.34.3
convex==0.7.0